==> source_code/01_collect_data.ipynb <==
<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Data collecting
</div>

<div style= "color: white">
<center>
Group 10

Quách Tấn Dũng - 21127247

Phạm Trần Minh Duy - 21127257

Đặng Quốc Thái - 21127545
</center>
</div>

import pandas as pd
import requests
import json

### Collect data
-   Data get from: 
    -   https://data.apps.fao.org/
    -   https://www.fao.org/faostat/en/#data

### What data we will collect:

domain_code = pd.read_csv("../data/external/domain_code.csv")
domain_code

area_code = 237 # Vietnam area code according to FAO

contents = []

for code in domain_code['Domain Code']:
    contents.append(requests.get(f"https://fenixservices.fao.org/faostat/api/v1/en/data/{code}?area={area_code}"))

-   what type of content we just collected?

print(contents[0].headers['Content-Type'])

-   So the content that we get from that URL is a JSON style content so we convert it to JSON format and save to a Dataframe

raw_dfs = []
for con in contents:
    raw_dfs.append(pd.DataFrame(con.json()['data']))

for df in raw_dfs:
    display(df.head(3))

-   Let's test if we have at least 1000 observations and 5 fields

for df in raw_dfs:
    print(df.shape)

-   Look good enough!

### Save collected data to file
-   Now we save the DataFrame to a csv file

for df in raw_dfs:
    domain_code = df['Domain Code'].values[0]
    df.to_csv(f"../data/raw/{domain_code}.csv", index= False)
