==> source_code/02_preprocess_data.ipynb <==
<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Preprocess Data
</div>

import pandas as pd

## Read raw data that we have collected

annual_pop_df = pd.read_csv("../data/raw/OA.csv")
agri_employments_df = pd.read_csv("../data/raw/OEA.csv")
rural_employments_df = pd.read_csv("../data/raw/OER.csv")
values_df = pd.read_csv("../data/raw/QV.csv")
products_df = pd.read_csv("../data/raw/QCL.csv")

display(annual_pop_df.head(3))
display(agri_employments_df.head(3))
display(rural_employments_df.head(3))
display(values_df.head(3))
display(products_df.head(3))

## Determine the features that have a large number of missing values

for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    percent_missing = df.isnull().sum() * 100 / len(df)
    missing_value_df = pd.DataFrame({'column_name': df.columns,
                                    'percent_missing': percent_missing})
    
    miss = (missing_value_df['column_name'][missing_value_df['percent_missing'].values >= 75]).values
    print(f"{df['Domain'].values[0]}, features with large missing value: {miss} (>= 75%)")

-   `annual_pop_df` and `products_df` have `Note` as the features that have a large number of missing values, but for now, we will keep it

## Check data type of each columns

for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    df.info()

-   For all of the above dataframe: `Area Code`, `Element Code`, `Item Code`, `Year` , `Indicator Code` have numeric type. However, their magnitude does not have a significance. And, `Year` actually represents a period instead of a number, `Area Code`, `Element Code`, `Item Code`, `Indicator Code` are just a category. Thus, they can be convert to categorical type.

for col in annual_pop_df.drop(columns= 'Value').columns:
    annual_pop_df[col] = annual_pop_df[col].astype(str)

for col in agri_employments_df.drop(columns= 'Value').columns:
    agri_employments_df[col] = agri_employments_df[col].astype(str)

for col in rural_employments_df.drop(columns= 'Value').columns:
    rural_employments_df[col] = rural_employments_df[col].astype(str)

for col in values_df.drop(columns= 'Value').columns:
    values_df[col] = values_df[col].astype(str)

for col in products_df.drop(columns= 'Value').columns:
    products_df[col] = products_df[col].astype(str)

-   Let check one last time

for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    df.info()

## Clear duplicates columns:
-   There is `Domain Code` and `Domain` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Area Code` and `Area` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Element Code` and `Element` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Indicator Code` and `Indicator` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Item Code` and `Item` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Year Code` and `Year` columns, they have the same values, so we drop the `Year Code`.
-   There is `Source Code` and `Source` columns, they have the same meaning, we will create a new dataframe to store it.
-   There is `Flag` and `Flag Description` columns, they have the same meaning, we will create a new dataframe to store it.

- Domain Code: 

domain = []
domain_code = []
for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    codes = df['Domain Code'].unique()
    for c in codes:
        dom = df['Domain'][df['Domain Code'] == c].values[0]
        domain_code.append(c)
        domain.append(dom)


domain_df = pd.DataFrame(list(zip(domain_code, domain)), columns= ['Domain Code', 'Domain'])
domain_df.to_csv("../data/external/domain_code.csv", index= False)
domain_df.head()

-   Area Code

area = []
area_code = []
for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    codes = df['Area Code'].unique()
    for c in codes: 
        are = df['Area'][df['Area Code'] == c].values[0]
        area_code.append(c)
        area.append(are)


area_df = pd.DataFrame(list(zip(area_code, area)), columns= ['Area Code', 'Area'])
area_df = area_df.drop_duplicates()
area_df.to_csv("../data/external/area_code.csv", index= False)
area_df.head()

-   Element Code

element = []
element_code = []
for df in [annual_pop_df, values_df, products_df, agri_employments_df, rural_employments_df]:
    codes = df['Element Code'].unique()
    for c in codes: 
        ele = df['Element'][df['Element Code'] == c].values[0]
        element_code.append(c)
        element.append(ele + f" ({(df[df['Element Code'] == c]['Unit'].unique()[0])})")


element_df = pd.DataFrame(list(zip(element_code, element)), columns= ['Element Code', 'Element'])
element_df = element_df.drop_duplicates()
element_df.to_csv("../data/external/element_code.csv", index= False)
element_df.head()

-   Indicator Code

indicator = []
indicator_code = []
for df in [agri_employments_df, rural_employments_df]:
    codes = df['Indicator Code'].unique()
    for c in codes: 
        ind = df['Indicator'][df['Indicator Code'] == c].values[0]
        indicator_code.append(c)
        indicator.append(ind)


indicator_df = pd.DataFrame(list(zip(indicator_code, indicator)), columns= ['Indicator Code', 'Indicator'])
indicator_df = indicator_df.drop_duplicates()
indicator_df.to_csv("../data/external/indicator_code.csv", index= False)
indicator_df.head()

-   Item Code

item = []
item_code = []
for df in [annual_pop_df, values_df, products_df]:
    codes = df['Item Code'].unique()
    for c in codes: 
        ite = df['Item'][df['Item Code'] == c].values[0]
        item_code.append(c)
        item.append(ite)


item_df = pd.DataFrame(list(zip(item_code, item)), columns= ['Item Code', 'Item'])
item_df = item_df.drop_duplicates()
item_df.to_csv("../data/external/item_code.csv", index= False)
item_df.head()

- Year Code

annual_pop_df = annual_pop_df.drop(columns='Year Code')
agri_employments_df = agri_employments_df.drop(columns='Year Code')
rural_employments_df = rural_employments_df.drop(columns='Year Code')
values_df = values_df.drop(columns='Year Code')
products_df = products_df.drop(columns='Year Code')

-   Source Code

source = []
source_code = []
for df in [agri_employments_df, rural_employments_df]:
    codes = df['Source Code'].unique()
    for c in codes: 
        ind = df['Source'][df['Source Code'] == c].values[0]
        source_code.append(c)
        source.append(ind)


source_df = pd.DataFrame(list(zip(source_code, source)), columns= ['Source Code', 'Source'])
source_df = source_df.drop_duplicates()
source_df.to_csv("../data/external/source_code.csv", index= False)
source_df.head()

-   Flag Description

flag = []
flag_des = []
for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    flags = df['Flag'].unique()
    for f in flags:
        des = df['Flag Description'][df['Flag'] == f].values[0]
        flag.append(f)
        flag_des.append(des)

flag_df = pd.DataFrame(list(zip(flag, flag_des)), columns= ['Flag', 'Flag Description'])
flag_df = flag_df.drop_duplicates()
flag_df.to_csv("../data/external/flag_des.csv", index= False)
flag_df.head()

-   Now we drop all duplicate columns

annual_pop_df = annual_pop_df.drop(columns=['Flag Description', 'Area', 'Domain', 'Item', 'Element'])
agri_employments_df = agri_employments_df.drop(columns=['Flag Description', 'Area', 'Domain', 'Source', 'Indicator'])
rural_employments_df = rural_employments_df.drop(columns=['Flag Description', 'Area Code', 'Domain', 'Source', 'Indicator'])
values_df = values_df.drop(columns=['Flag Description', 'Area', 'Domain', 'Item', 'Element'])
products_df = products_df.drop(columns=['Flag Description', 'Area', 'Domain', 'Item', 'Element'])

display(annual_pop_df.head(3))
display(agri_employments_df.head(3))
display(rural_employments_df.head(3))
display(values_df.head(3))
display(products_df.head(3))

## Check for duplicate rows

for df in [annual_pop_df, agri_employments_df, rural_employments_df, values_df, products_df]:
    index = agri_employments_df.index
    detectDupSeries = index.duplicated(keep='first')
    num_duplicated_rows = detectDupSeries.sum()
    if num_duplicated_rows == 0:
        print(f"{df['Domain Code'].values[0]} have no duplicated line.!")
    else:
        if num_duplicated_rows > 1:
            ext = "lines"
        else:
            ext = "line"
        print(f"{df['Domain Code'].values[0]} {num_duplicated_rows} duplicated " + ext + ". Please de-deduplicate your raw data.!")

### Save processed data to a new csv file

annual_pop_df.to_csv("../data/processed/preprocess_OA.csv", index= False)
agri_employments_df.to_csv("../data/processed/preprocess_OEA.csv", index= False)
rural_employments_df.to_csv("../data/processed/preprocess_OER.csv", index= False)
values_df.to_csv("../data/processed/preprocess_QV.csv", index= False)
products_df.to_csv("../data/processed/preprocess_QCL.csv", index= False)
