==> source_code/02.1_explain_data.ipynb <==
<div style="text-align: center; background-color: #5A96E3; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Explain Data
</div>

import pandas as pd

oa=pd.read_csv("../data/raw/OA.csv")
# oa

`COLUMNS:`
- `Domain Code:` A numerical identifier for a particular domain or field of study, in this case, `OA`.
- `Domain:` The specific name or area represented by the domain code, in this case, `Annual Population`.
- `Area Code:` A numerical identifier for a country, region, or specific geographical area using the FAO code system.
- `Area:` The name or designation of the corresponding territory identified by the area code.
- `Element Code:` A numerical identifier for a distinct element or factor within a dataset.
- `Element:` The name or description corresponding to the element code.
- `Item Code:` A numerical identifier for a particular item or category within a dataset.
- `Item:` The name or category represented by the item code.
- `Year Code:` A numerical identifier for a specific year within a timeline.
- `Year:` The calendar year corresponding to the year code.
- `Unit:` The multiplier value, used to multiply the "value" column in order to derive the actual value.
- `Value:` The numerical figure representing the population count for a specified area.
- `Flag:` The code for Observation Status to be associated to secondary data disseminated by FAO.
- `Flag Description:` A detailed explanation or description outlining the significance or implications of the flag.
- `Note:` notes if any

`ROWS:` The rows contain specific details pertaining to an element, showcasing cyclical data structured through out years as follows:
- Total Population - Both sexes
- Total Population - Male
- Total Population - Female
- Rural population	
- Urban population	

oea=pd.read_csv("../data/raw/OEA.csv")
# oea

`COLUMNS:`
- `Domain Code:` A numerical identifier for a particular domain or field of study.
- `Domain:` The specific name or area represented by the domain code.
- `Area Code:` A numerical identifier for a country, region, or specific geographical area using the FAO code system.
- `Area:` The name or designation of the corresponding territory identified by the area code.
- `Indicator Code:` A numerical label assigned to specific data points within dataset for classification and identification purposes.
- `Indicator:` A factor used to represent and assess a particular aspect or trend within dataset, providing insights into subjects or domains.
- `Source code:` A numerical identifier representing the specific source or origin of the data within the dataset.
- `Source:` The name of the source or provider of the data inditified by the source code.
- `Sex Code:` A numerical values representing gender or sex categories.
- `Sex:` Provides the descriptive labels corresponding to the sex codes.
- `Element Code:` A numerical identifier for a distinct element or factor within a dataset.
- `Element:` The name or description corresponding to the element code.
- `Year Code:` A numerical identifier for a specific year within a timeline.
- `Year:` The calendar year corresponding to the year code.
- `Unit:` The multiplier value, used to multiply the "value" column in order to derive the actual value.
- `Value:` The numerical figure representing the population count for a specified area.
- `Flag:` The code for Observation Status to be associated to secondary data disseminated by FAO.
- `Flag Description:` A detailed explanation or description outlining the significance or implications of the flag.
- `Note:` notes if any.

`ROWS:` The rows present information regarding an element, demonstrating a sequential data structure combines 'Indicator' and 'Sex' through out years as follows:
- Indicator:
    - Employment in agriculture, forestry and fishing by age:
        + Total (15+)
        + 15 to 24
        + 25 to 54
        + 55 to 64
        + 65+
    - Employment in agriculture, forestry and fishing by status in employment:
        + Total
        + Employees
        + Self-employed
        + Workers not classified
    - Employment in agriculture, forestry and fishing by sub-sectors:
        + Crop and animal production, hunting and related service activities
        + Forestry and logging
        + Fishing and aquaculture
    - Mean weekly hours actually worked employed person in agriculture, forestry and fishing
    - Mean weekly hours actually worked per employee in agriculture, forestry and fishing
    - Employment in agriculture, forestry and fishing - ILO Modelled Estimates
    - Share of employment in agriculture, forestry and fishing by sub-sectors:
        + Crop and animal production, hunting and related service activities
        + Forestry and logging
        + Fishing and aquaculture
    - Share of employment in agriculture, forestry and fishing in total employment
    - Share of female in total employment in agriculture, forestry and fishing
    - Share of employees in agriculture, forestry and fishing in total employees
    - Share of female employees in total employment in agriculture, forestry and fishing
    - Agriculture value added per worker (constant 2015 US$)
    - Share of employment in agriculture, forestry and fishing in total employment - ILO Modelled Estimates
- Sex:
    - Total
    - Male
    - Female

oer=pd.read_csv("../data/raw/OER.csv")
# oer

`COLUMNS:`
- `Domain Code:` A numerical identifier for a particular domain or field of study.
- `Domain:` The specific name or area represented by the domain code.
- `Area Code:` A numerical identifier for a country, region, or specific geographical area using the FAO code system.
- `Area:` The name or designation of the corresponding territory identified by the area code.
- `Indicator Code:` A numerical label assigned to specific data points within dataset for classification and identification purposes.
- `Indicator:` A factor used to represent and assess a particular aspect or trend within dataset, providing insights into subjects or domains.
- `Source code:` A numerical identifier representing the specific source or origin of the data within the dataset.
- `Source:` The name of the source or provider of the data inditified by the source code.
- `Sex Code:` A numerical values representing gender or sex categories.
- `Sex:` Provides the descriptive labels corresponding to the sex codes.
- `Element Code:` A numerical identifier for a distinct element or factor within a dataset.
- `Element:` The name or description corresponding to the element code.
- `Year Code:` A numerical identifier for a specific year within a timeline.
- `Year:` The calendar year corresponding to the year code.
- `Unit:` The multiplier value, used to multiply the "value" column in order to derive the actual value.
- `Value:` The numerical figure representing the population count for a specified area.
- `Flag:` The code for Observation Status to be associated to secondary data disseminated by FAO.
- `Flag Description:` A detailed explanation or description outlining the significance or implications of the flag.
- `Note:` notes if any.

`ROWS:` The rows present information regarding an element, demonstrating a sequential data structure combines 'Indicator' and 'Sex' through out years as follows:
- Indicator:
    - Employment by age in rural areas:
        + Total (15+)
        + 15 to 24
        + 25 to 54
        + 55 to 64
        + 65+
    - Employment by status in employment in rural areas:
        + Total
        + Employees
        + Self-employed
        + Workers not classified
    - Labour force participant rate by age in rural areas:
        + Total (15+)
        + 15 to 24
        + 25 to 54
        + 55 to 64
        + 65+
    - Employment-to-population ratio by age in rural areas:
        + Total (15+)
        + 15 to 24
        + 25 to 54
        + 55 to 64
        + 65+
    - Share of employment by status in employment in rural areas:
        + Total
        + Employees
        + Self-employed
        + Workers not classified
- Sex:
    - Total
    - Male
    - Female

qcl=pd.read_csv("../data/raw/QCL.csv")
qcl

`COLUMNS:`
- `Domain Code:` A numerical identifier for a particular domain or field of study, in this case, `OA`.
- `Domain:` The specific name or area represented by the domain code, in this case, `Annual Population`.
- `Area Code:` A numerical identifier for a country, region, or specific geographical area using the FAO code system.
- `Area:` The name or designation of the corresponding territory identified by the area code.
- `Element Code:` A numerical identifier for a distinct element or factor within a dataset.
- `Element:` The name or description corresponding to the element code.
- `Item Code:` A numerical identifier for a particular item or category within a dataset.
- `Item:` The name or category represented by the item code.
- `Year Code:` A numerical identifier for a specific year within a timeline.
- `Year:` The calendar year corresponding to the year code.
- `Unit:` The multiplier value, used to multiply the "value" column in order to derive the actual value.
- `Value:` The numerical figure representing the population count for a specified area.
- `Flag:` The code for Observation Status to be associated to secondary data disseminated by FAO.
- `Flag Description:` A detailed explanation or description outlining the significance or implications of the flag.
- `Note:` notes if any

`ROWS:` The rows present information regarding an element, demonstrating a sequential data structure combines 'Element' and 'Item' through out years as follows:
- Elements:
   - Area harvested
   - Yield: represents the quantity of a type of crops harvested per an unit of area or an amount product gained per animal
      + 100 g/ha
      + 100 mg/An: hundred Milligrams per animal
      + No/An: Number per animal
      + Yield/Carcass Weight (100 g/An)
      + Yield/Carcass Weight (0.1 g/An)
      + 100 g
      + 100 g/An: hundred Grams per animal
   - Production: refers to the quantity of crops or livestock yielded, indicating the output of agricultural goods.
      + Ton
      + 1000 No: thousand Number
   - Stocks: represents the quantity of crops or livestock harvested within a specific timeframe
   - Producing Animals/Slaughtered:
      + An: Animal
      + 1000 An: thousand Animals
- Items:
   - Crops, primary:
      + Abaca, manila hemp, raw 
      + Agave fibres, raw, n.e.c. 
      + Almonds, in shell 
      + Anise, badian, coriander, cumin, caraway, fennel and juniper berries, raw 
      + Apples 
      + Apricots 
      + Areca nuts 
      + Artichokes 
      + Asparagus 
      + Avocados 
      + Balata, gutta-percha, guayule, chicle and similar natural gums in primary forms or in plates, sheets or strip 
      + Bambara beans, dry 
      + Bananas 
      + Barley 
      + Beans, dry 
      + Blueberries 
      + Brazil nuts, in shell 
      + Broad beans and horse beans, dry 
      + Broad beans and horse beans, green 
      + Buckwheat 
      + Cabbages 
      + Canary seed 
      + Cantaloupes and other melons 
      + Carrots and turnips 
      + Cashew nuts, in shell 
      + Cashewapple 
      + Cassava leaves 
      + Cassava, fresh 
      + Castor oil seeds 
      + Cauliflowers and broccoli 
      + Cereals n.e.c. 
      + Cherries 
      + Chestnuts, in shell 
      + Chick peas, dry 
      + Chicory roots 
      + Chillies and peppers, dry (Capsicum spp., Pimenta spp.), raw 
      + Chillies and peppers, green (Capsicum spp. and Pimenta spp.) 
      + Cinnamon and cinnamon-tree flowers, raw 
      + Cloves (whole stems), raw 
      + Cocoa beans 
      + Coconuts, in shell 
      + Coffee, green 
      + Coir, raw 
      + Cow peas, dry 
      + Cranberries 
      + Cucumbers and gherkins 
      + Currants 
      + Dates 
      + Edible roots and tubers with high starch or inulin content, n.e.c., fresh 
      + Eggplants (aubergines) 
      + Figs 
      + Flax, processed but not spun 
      + Fonio 
      + Ginger, raw 
      + Gooseberries 
      + Grapes 
      + Green corn (maize) 
      + Green garlic 
      + Groundnuts, excluding shelled 
      + Hazelnuts, in shell 
      + Hempseed 
      + Hop cones 
      + Jojoba seeds 
      + Jute, raw or retted 
      + Kapok fruit 
      + Karite nuts (sheanuts) 
      + Kenaf, and other textile bast fibres, raw or retted 
      + Kiwi fruit 
      + Kola nuts 
      + Leeks and other alliaceous vegetables 
      + Lemons and limes 
      + Lentils, dry 
      + Lettuce and chicory 
      + Linseed 
      + Locust beans (carobs) 
      + Lupins 
      + Maize (corn) 
      + Mangoes, guavas and mangosteens
      + Maté leaves 
      + Melonseed 
      + Millet 
      + Mixed grain 
      + Mushrooms and truffles 
      + Mustard seed 
      + Natural rubber in primary forms 
      + Nutmeg, mace, cardamoms, raw 
      + Oats 
      + Oil palm fruit 
      + Okra 
      + Olives 
      + Onions and shallots, dry (excluding dehydrated) 
      + Onions and shallots, green 
      + Oranges 
      + Other beans, green 
      + Other berries and fruits of the genus vaccinium n.e.c. 
      + Other citrus fruit, n.e.c. 
      + Other fibre crops, raw, n.e.c. 
      + Other fruits, n.e.c. 
      + Other nuts (excluding wild edible nuts and groundnuts), in shell, n.e.c. 
      + Other oil seeds, n.e.c. 
      + Other pome fruits 
      + Other pulses n.e.c. 
      + Other stimulant, spice and aromatic crops, n.e.c. 
      + Other stone fruits 
      + Other sugar crops n.e.c. 
      + Other tropical fruits, n.e.c. 
      + Other vegetables, fresh n.e.c. 
      + Papayas 
      + Peaches and nectarines 
      + Pears 
      + Peas, dry 
      + Peas, green 
      + Pepper (Piper spp.), raw 
      + Peppermint, spearmint 
      + Persimmons 
      + Pigeon peas, dry 
      + Pineapples 
      + Pistachios, in shell 
      + Plantains and cooking bananas 
      + Plums and sloes 
      + Pomelos and grapefruits 
      + Poppy seed 
      + Potatoes 
      + Pumpkins, squash and gourds 
      + Pyrethrum, dried flowers 
      + Quinces 
      + Quinoa 
      + Ramie, raw or retted 
      + Rape or colza seed 
      + Raspberries 
      + Rice 
      + Rye 
      + Safflower seed 
      + Seed cotton, unginned 
      + Sesame seed 
      + Sisal, raw 
      + Sorghum 
      + Sour cherries 
      + Soya beans 
      + Spinach 
      + Strawberries 
      + String beans 
      + Sugar beet 
      + Sugar cane 
      + Sunflower seed 
      + Sweet potatoes 
      + Tallowtree seeds 
      + Tangerines, mandarins, clementines 
      + Taro 
      + Tea leaves 
      + Tomatoes 
      + Triticale 
      + True hemp, raw or retted 
      + Tung nuts 
      + Unmanufactured tobacco 
      + Vanilla, raw 
      + Vetches 
      + Walnuts, in shell 
      + Watermelons 
      + Wheat 
      + Yams 
      + Yautia
   - Crops Processed:
      + Beer of barley, malted 
      + Coconut oil 
      + Cotton lint, ginned 
      + Cotton seed 
      + Cottonseed oil 
      + Green tea (not fermented), black tea (fermented) and partly fermented tea, in immediate packings of a content not exceeding 3 kg
      + Groundnut oil 
      + Kapok fibre, raw 
      + Kapokseed in shell 
      + Margarine and shortening 
      + Molasses 
      + Oil of linseed 
      + Oil of maize 
      + Oil of palm kernel 
      + Oil of sesame seed 
      + Olive oil 
      + Palm kernels 
      + Palm oil 
      + Rapeseed or canola oil, crude 
      + Raw cane or beet sugar (centrifugal only) 
      + Safflower-seed oil, crude 
      + Soya bean oil 
      + Sunflower-seed oil, crude 
      + Wine
   - Livestock primary:
      + Beeswax 
      + Buffalo fat, unrendered 
      + Cattle fat, unrendered 
      + Edible offal of buffalo, fresh, chilled or frozen 
      + Edible offal of cattle, fresh, chilled or frozen 
      + Edible offal of goat, fresh, chilled or frozen 
      + Edible offal of pigs, fresh, chilled or frozen 
      + Edible offal of sheep, fresh, chilled or frozen 
      + Edible offals of camels and other camelids, fresh, chilled or frozen 
      + Edible offals of horses and other equines, fresh, chilled or frozen 
      + Eggs from other birds in shell, fresh, n.e.c.
      + Fat of camels 
      + Fat of pigs 
      + Game meat, fresh, chilled or frozen 
      + Goat fat, unrendered 
      + Hen eggs in shell, fresh 
      + Horse meat, fresh or chilled 
      + Meat of asses, fresh or chilled 
      + Meat of buffalo, fresh or chilled 
      + Meat of camels, fresh or chilled 
      + Meat of cattle with the bone, fresh or chilled 
      + Meat of chickens, fresh or chilled 
      + Meat of ducks, fresh or chilled 
      + Meat of geese, fresh or chilled 
      + Meat of goat, fresh or chilled 
      + Meat of mules, fresh or chilled 
      + Meat of other domestic camelids, fresh or chilled 
      + Meat of other domestic rodents, fresh or chilled 
      + Meat of pig with the bone, fresh or chilled 
      + Meat of pigeons and other birds n.e.c., fresh, chilled or frozen 
      + Meat of rabbits and hares, fresh or chilled 
      + Meat of sheep, fresh or chilled 
      + Meat of turkeys, fresh or chilled 
      + Natural honey 
      + Other birds 
      + Other meat n.e.c. (excluding mammals), fresh, chilled or frozen 
      + Raw hides and skins of buffaloes 
      + Raw hides and skins of cattle 
      + Raw hides and skins of goats or kids 
      + Raw hides and skins of sheep or lambs 
      + Raw milk of buffalo 
      + Raw milk of camel 
      + Raw milk of cattle 
      + Raw milk of goats 
      + Raw milk of sheep 
      + Sheep fat, unrendered 
      + Shorn wool, greasy, including fleece-washed shorn wool 
      + Silk-worm cocoons suitable for reeling 
      + Snails, fresh, chilled, frozen, dried, salted or in brine, except sea snails
   - Livestock processed:
      + Butter and ghee of sheep milk 
      + Butter of buffalo milk 
      + Butter of cow milk 
      + Butter of goat milk 
      + Buttermilk, dry 
      + Cheese from milk of buffalo, fresh or processed 
      + Cheese from milk of goats, fresh or processed 
      + Cheese from milk of sheep, fresh or processed 
      + Cheese from skimmed cow milk 
      + Cheese from whole cow milk 
      + Cream, fresh 
      + Ghee from buffalo milk 
      + Ghee from cow milk 
      + Livestock processed > (List) 
      + Pig fat, rendered 
      + Raw silk (not thrown) 
      + Skim milk and whey powder 
      + Skim milk of cows 
      + Skim milk, condensed 
      + Skim milk, evaporated 
      + Tallow 
      + Whey, condensed 
      + Whey, dry 
      + Whole milk powder 
      + Whole milk, condensed 
      + Whole milk, evaporated 
      + Yoghurt
   - Live animals:
      + Asses 
      + Bees 
      + Buffalo 
      + Camels 
      + Cattle 
      + Chickens 
      + Ducks 
      + Geese 
      + Goats 
      + Horses 
      + Mules and hinnies 
      + Other birds 
      + Other camelids 
      + Other rodents 
      + Rabbits and hares 
      + Sheep 
      + Swine / pigs 
      + Turkeys 

qv=pd.read_csv("../data/raw/QV.csv")
qv

`COLUMNS:`
- `Domain Code:` A numerical identifier for a particular domain or field of study, in this case, `OA`.
- `Domain:` The specific name or area represented by the domain code, in this case, `Annual Population`.
- `Area Code:` A numerical identifier for a country, region, or specific geographical area using the FAO code system.
- `Area:` The name or designation of the corresponding territory identified by the area code.
- `Element Code:` A numerical identifier for a distinct element or factor within a dataset.
- `Element:` The name or description corresponding to the element code.
- `Item Code:` A numerical identifier for a particular item or category within a dataset.
- `Item:` The name or category represented by the item code.
- `Year Code:` A numerical identifier for a specific year within a timeline.
- `Year:` The calendar year corresponding to the year code.
- `Unit:` The multiplier value, used to multiply the "value" column in order to derive the actual value.
- `Value:` The numerical figure representing the population count for a specified area.
- `Flag:` The code for Observation Status to be associated to secondary data disseminated by FAO.
- `Flag Description:` A detailed explanation or description outlining the significance or implications of the flag.
- `Note:` notes if any

`ROWS:` The rows present information regarding an element, demonstrating a sequential data structure combines 'Element' and 'Item' through out years as follows:
- Elements:
    - Gross Production Value (constant 2014-2016 thousand I$): This represents the total value of goods and services produced within an economy, adjusted for inflation to maintain the equivalent purchasing power of the dollar from the 2014-2016 period. It's measured in thousands of international dollars (I$) and is utilized to compare economic output across different years while accounting for changes in price levels.
    - Gross Production Value (constant 2014-2016 thousand SLC): Similar to the previous description, this measures the total value of goods and services produced, adjusted for inflation to maintain the equivalent purchasing power of the currency for the 2014-2016 period. It's expressed in thousands of standard local currency units (SLC) and allows for comparison of economic output over time, adjusted for changes in the price levels of the reference period.
    - Gross Production Value (constant 2014-2016 thousand US$): This denotes the total value of goods and services produced, adjusted for inflation to maintain the equivalent purchasing power of the United States dollar from the 2014-2016 period. It's measured in thousands of US dollars (US$) and facilitates comparison of economic output across different years while accounting for changes in price levels.
    - Gross Production Value (current thousand SLC): Represents the current, unadjusted value of goods and services produced, measured in thousands of standard local currency units (SLC). It reflects the current monetary worth of production without adjusting for inflation, providing a snapshot of the current economic output.
    - Gross Production Value (current thousand US$): Denotes the current, unadjusted value of goods and services produced, measured in thousands of United States dollars (US$). Similar to the previous entry, it shows the current monetary worth of production without adjusting for inflation, providing a snapshot of the current economic output in US dollars.
- Items:
    - Abaca, manila hemp, raw 
    - Agave fibres, raw, n.e.c. 
    - Almonds, in shell 
    - Anise, badian, coriander, cumin, caraway, fennel and juniper berries, raw 
    - Apples 
    - Apricots 
    - Areca nuts 
    - Artichokes 
    - Asparagus 
    - Avocados 
    - Balata, gutta-percha, guayule, chicle and similar natural gums in primary forms or in plates, sheets or strip 
    - Bambara beans, dry 
    - Bananas 
    - Barley 
    - Beans, dry 
    - Beeswax 
    - Blueberries 
    - Brazil nuts, in shell 
    - Broad beans and horse beans, dry 
    - Broad beans and horse beans, green 
    - Buckwheat 
    - Buffalo fat, unrendered 
    - Cabbages 
    - Canary seed 
    - Cantaloupes and other melons 
    - Carrots and turnips 
    - Cashew nuts, in shell 
    - Cashewapple 
    - Cassava, fresh 
    - Castor oil seeds 
    - Cattle fat, unrendered 
    - Cauliflowers and broccoli 
    - Cereals n.e.c. 
    - Cherries 
    - Chestnuts, in shell 
    - Chick peas, dry 
    - Chicory roots 
    - Chillies and peppers, dry (Capsicum spp., Pimenta spp.), raw 
    - Chillies and peppers, green (Capsicum spp. and Pimenta spp.) 
    - Cinnamon and cinnamon-tree flowers, raw 
    - Cloves (whole stems), raw 
    - Cocoa beans 
    - Coconuts, in shell 
    - Coffee, green 
    - Cow peas, dry 
    - Cranberries 
    - Cucumbers and gherkins 
    - Currants 
    - Dates 
    - Edible offal of buffalo, fresh, chilled or frozen 
    - Edible offal of cattle, fresh, chilled or frozen 
    - Edible offal of goat, fresh, chilled or frozen 
    - Edible offal of pigs, fresh, chilled or frozen 
    - Edible offal of sheep, fresh, chilled or frozen 
    - Edible offals of camels and other camelids, fresh, chilled or frozen 
    - Edible offals of horses and other equines, fresh, chilled or frozen 
    - Edible roots and tubers with high starch or inulin content, n.e.c., fresh 
    - Eggplants (aubergines) 
    - Eggs from other birds in shell, fresh, n.e.c. 
    - Fat of camels 
    - Fat of pigs 
    - Figs 
    - Flax, processed but not spun 
    - Fonio 
    - Game meat, fresh, chilled or frozen 
    - Ginger, raw 
    - Goat fat, unrendered 
    - Gooseberries 
    - Grapes 
    - Green corn (maize) 
    - Green garlic 
    - Groundnuts, excluding shelled 
    - Hazelnuts, in shell 
    - Hempseed 
    - Hen eggs in shell, fresh 
    - Hop cones 
    - Horse meat, fresh or chilled 
    - Horse meat, fresh or chilled (indigenous) 
    - Jojoba seeds 
    - Jute, raw or retted 
    - Kapok fruit 
    - Karite nuts (sheanuts) 
    - Kenaf, and other textile bast fibres, raw or retted  Kiwi fruit 
    - Kola nuts 
    - Leeks and other alliaceous vegetables 
    - Lemons and limes 
    - Lentils, dry 
    - Lettuce and chicory 
    - Linseed 
    - Locust beans (carobs) 
    - Lupins 
    - Maize (corn) 
    - Mangoes, guavas and mangosteens 
    - Maté leaves 
    - Meat of asses, fresh or chilled 
    - Meat of asses, fresh or chilled (indigenous) 
    - Meat of buffalo, fresh or chilled 
    - Meat of buffalo, fresh or chilled (indigenous) 
    - Meat of camels, fresh or chilled 
    - Meat of camels, fresh or chilled (indigenous) 
    - Meat of cattle with the bone, fresh or chilled 
    - Meat of cattle with the bone, fresh or chilled (indigenous) 
    - Meat of chickens, fresh or chilled 
    - Meat of chickens, fresh or chilled (indigenous) 
    - Meat of ducks, fresh or chilled 
    - Meat of ducks, fresh or chilled (indigenous) 
    - Meat of geese, fresh or chilled 
    - Meat of geese, fresh or chilled (indigenous) 
    - Meat of goat, fresh or chilled 
    - Meat of goat, fresh or chilled (indigenous) 
    - Meat of mules, fresh or chilled 
    - Meat of mules, fresh or chilled (indigenous) 
    - Meat of other domestic camelids, fresh or chilled 
    - Meat of other domestic camelids, fresh or chilled (indigenous) 
    - Meat of other domestic rodents, fresh or chilled 
    - Meat of other domestic rodents, fresh or chilled (indigenous) 
    - Meat of pig with the bone, fresh or chilled 
    - Meat of pig with the bone, fresh or chilled (indigenous) 
    - Meat of pigeons and other birds n.e.c., fresh, chilled or frozen 
    - Meat of pigeons and other birds n.e.c., fresh, chilled or frozen (indigenous) 
    - Meat of rabbits and hares, fresh or chilled 
    - Meat of rabbits and hares, fresh or chilled (indigenous) 
    - Meat of sheep, fresh or chilled 
    - Meat of sheep, fresh or chilled (indigenous) 
    - Meat of turkeys, fresh or chilled 
    - Meat of turkeys, fresh or chilled (indigenous) 
    - Melonseed 
    - Millet 
    - Mixed grain 
    - Mushrooms and truffles 
    - Mustard seed 
    - Natural honey 
    - Natural rubber in primary forms 
    - Nutmeg, mace, cardamoms, raw 
    - Oats 
    - Oil palm fruit 
    - Okra 
    - Olives 
    - Onions and shallots, dry (excluding dehydrated) 
    - Onions and shallots, green 
    - Oranges 
    - Other beans, green 
    - Other berries and fruits of the genus vaccinium n.e.c. 
    - Other citrus fruit, n.e.c. 
    - Other fibre crops, raw, n.e.c. 
    - Other fruits, n.e.c. 
    - Other meat n.e.c. (excluding mammals), fresh, chilled or frozen 
    - Other nuts (excluding wild edible nuts and groundnuts), in shell, n.e.c. 
    - Other oil seeds, n.e.c. 
    - Other pome fruits 
    - Other pulses n.e.c. 
    - Other stimulant, spice and aromatic crops, n.e.c. 
    - Other stone fruits 
    - Other sugar crops n.e.c. 
    - Other tropical fruits, n.e.c. 
    - Other vegetables, fresh n.e.c. 
    - Papayas 
    - Peaches and nectarines 
    - Pears 
    - Peas, dry 
    - Peas, green 
    - Pepper (Piper spp.), raw 
    - Peppermint, spearmint 
    - Persimmons 
    - Pigeon peas, dry 
    - Pineapples 
    - Pistachios, in shell 
    - Plantains and cooking bananas 
    - Plums and sloes 
    - Pomelos and grapefruits 
    - Poppy seed 
    - Potatoes 
    - Pumpkins, squash and gourds 
    - Pyrethrum, dried flowers 
    - Quinces 
    - Quinoa 
    - Ramie, raw or retted 
    - Rape or colza seed 
    - Raspberries 
    - Raw hides and skins of goats or kids 
    - Raw hides and skins of sheep or lambs 
    - Raw milk of buffalo 
    - Raw milk of camel 
    - Raw milk of cattle 
    - Raw milk of goats 
    - Raw milk of sheep 
    - Raw skins of swine 
    - Rice 
    - Rye 
    - Safflower seed 
    - Seed cotton, unginned 
    - Sesame seed 
    - Sheep fat, unrendered 
    - Shorn wool, greasy, including fleece-washed shorn wool 
    - Silk-worm cocoons suitable for reeling 
    - Sisal, raw 
    - Snails, fresh, chilled, frozen, dried, salted or in brine, except sea snails 
    - Sorghum 
    - Sour cherries 
    - Soya beans 
    - Spinach 
    - Strawberries 
    - String beans 
    - Sugar beet 
    - Sugar cane 
    - Sunflower seed 
    - Sweet potatoes 
    - Tallowtree seeds 
    - Tangerines, mandarins, clementines 
    - Taro 
    - Tea leaves 
    - Tomatoes 
    - Triticale 
    - True hemp, raw or retted 
    - Tung nuts 
    - Unmanufactured tobacco 
    - Vanilla, raw 
    - Vetches 
    - Walnuts, in shell 
    - Watermelons 
    - Wheat 
    - Yams 
    - Yautia
